+++ R/do_AffinityAnalysisPlot.R
@@ -56,7 +56,7 @@ do_AffinityAnalysisPlot <- function(sample,
                                     plot.subtitle.face = "plain",
                                     plot.caption.face = "italic",
                                     axis.title.face = "bold",
-                                    axis.text.face = "bold",
+                                    axis.text.face = "plain",
                                     legend.title.face = "bold",
                                     legend.text.face = "plain"){
   # Add lengthy error messages.
@@ -421,6 +421,9 @@ do_AffinityAnalysisPlot <- function(sample,
       if (counter == 1){
         ylab <- group
         xlab <- NULL
+        if (length(group.by) == counter){
+          xlab <- "Gene set"
+        }
       } else {
         xlab <- "Gene set"
         ylab <- group