+++ R/do_CopyNumberVariantPlot.R
@@ -51,7 +51,7 @@ do_CopyNumberVariantPlot <- function(sample,
                                      plot.subtitle.face = "plain",
                                      plot.caption.face = "italic",
                                      axis.title.face = "bold",
-                                     axis.text.face = "bold",
+                                     axis.text.face = "plain",
                                      legend.title.face = "bold",
                                      legend.text.face = "plain"){
   # Add lengthy error messages.