+++ R/do_CorrelationPlot.R
@@ -48,7 +48,7 @@ do_CorrelationPlot <- function(sample = NULL,
                                plot.subtitle.face = "plain",
                                plot.caption.face = "italic",
                                axis.title.face = "bold",
-                               axis.text.face = "bold",
+                               axis.text.face = "plain",
                                legend.title.face = "bold",
                                legend.text.face = "plain"){
   # Add lengthy error messages.