+++ R/do_DiffusionMapPlot.R
@@ -54,7 +54,7 @@ do_DiffusionMapPlot <- function(sample,
                                 plot.subtitle.face = "plain",
                                 plot.caption.face = "italic",
                                 axis.title.face = "bold",
-                                axis.text.face = "bold",
+                                axis.text.face = "plain",
                                 legend.title.face = "bold",
                                 legend.text.face = "plain"){
   # Add lengthy error messages.