+++ R/do_DotPlot.R
@@ -58,7 +58,7 @@ do_DotPlot <- function(sample,
                        plot.subtitle.face = "plain",
                        plot.caption.face = "italic",
                        axis.title.face = "bold",
-                       axis.text.face = "bold",
+                       axis.text.face = "plain",
                        legend.title.face = "bold",
                        legend.text.face = "plain"){
     # Add lengthy error messages.