+++ R/do_ExpressionHeatmap.R
@@ -48,7 +48,7 @@ do_ExpressionHeatmap <- function(sample,
                                  plot.subtitle.face = "plain",
                                  plot.caption.face = "italic",
                                  axis.title.face = "bold",
-                                 axis.text.face = "bold",
+                                 axis.text.face = "plain",
                                  legend.title.face = "bold",
                                  legend.text.face = "plain"){