+++ R/do_GeyserPlot.R
@@ -72,7 +72,7 @@ do_GeyserPlot <- function(sample,
                           plot.subtitle.face = "plain",
                           plot.caption.face = "italic",
                           axis.title.face = "bold",
-                          axis.text.face = "bold",
+                          axis.text.face = "plain",
                           legend.title.face = "bold",
                           legend.text.face = "plain"){
   # Add lengthy error messages.