+++ R/do_GroupedGOTermPlot.R
@@ -34,7 +34,7 @@ do_GroupedGOTermPlot <- function(genes,
                                  plot.subtitle.face = "plain",
                                  plot.caption.face = "italic",
                                  axis.title.face = "bold",
-                                 axis.text.face = "bold",
+                                 axis.text.face = "plain",
                                  legend.title.face = "bold",
                                  legend.text.face = "plain"){
   # Add lengthy error messages.