+++ R/do_GroupwiseDEPlot.R
@@ -44,7 +44,7 @@ do_GroupwiseDEPlot <- function(sample,
                                plot.subtitle.face = "plain",
                                plot.caption.face = "italic",
                                axis.title.face = "bold",
-                               axis.text.face = "bold",
+                               axis.text.face = "plain",
                                legend.title.face = "bold",
                                legend.text.face = "plain"){
   # Add lengthy error messages.