+++ R/do_LoadingsPlot.R
@@ -50,7 +50,7 @@ do_LoadingsPlot <- function(sample,
                             plot.subtitle.face = "plain",
                             plot.caption.face = "italic",
                             axis.title.face = "bold",
-                            axis.text.face = "bold",
+                            axis.text.face = "plain",
                             legend.title.face = "bold",
                             legend.text.face = "plain"){
   # Add lengthy error messages.