+++ R/do_RidgePlot.R
@@ -63,7 +63,7 @@ do_RidgePlot <- function(sample,
                          plot.subtitle.face = "plain",
                          plot.caption.face = "italic",
                          axis.title.face = "bold",
-                         axis.text.face = "bold",
+                         axis.text.face = "plain",
                          legend.title.face = "bold",
                          legend.text.face = "plain"){
   # Add lengthy error messages.