+++ R/do_TermEnrichmentPlot.R
@@ -39,7 +39,7 @@ do_TermEnrichmentPlot <- function(enriched_terms,
                                   plot.subtitle.face = "plain",
                                   plot.caption.face = "italic",
                                   axis.title.face = "bold",
-                                  axis.text.face = "bold",
+                                  axis.text.face = "plain",
                                   legend.title.face = "bold",
                                   legend.text.face = "plain"){
     # Add lengthy error messages.